Create our own lights
Qt3D's lights aren't compatible with KHR_lights_punctual for a few reasons:
1. Need a "range" field for point and spot lights
2. spot lights need an innerConeAngle and outerConeAngle. Inner is the bright spot
   in center, outer represents the edge of the spotlight.  There should be nice
   attenuation between them.
3. directional lights should have the rotation portion of their node's
   transform applied.
4. Attenuation for spot and point lights should be quadratic (1/distance^2).

This patch creates reimplements QPointLight, QSpotLight, QDirectionalLight
with the required properties and behavior.  If we inherit from QAbstractLight
everything works OK.

Also fixes a bug in the old shader where we were doing diffuse and
specular calculations when sDotN < 0 resulting in really strange
lighting effects.

Task-Id: KUE-575
Reviewed-on: https://kuesa-codereview.kdab.com/c/kuesa/kuesa/+/65

+++ src/core/gltf2importer/lightparser.cpp
@@ -96,7 +96,7 @@ bool Kuesa::GLTF2Import::LightParser::parse(const QJsonArray &lights, Kuesa::GLT
             light.outerConeAngleRadians = spotObject.value(KEY_OUTER_CONE_ANGLE).toDouble(light.outerConeAngleRadians);
             if (light.innerConeAngleRadians >= light.outerConeAngleRadians || light.innerConeAngleRadians < 0)
                 qCWarning(kuesa) << "KHR_lights_punctual SpotLight innerConeAngle must be between 0 and outerConeAngle. Invalid light: " << light.name;
-            if (light.outerConeAngleRadians > M_PI_2)
+            if (light.outerConeAngleRadians > static_cast<float>(M_PI_2))
                 qCWarning(kuesa) << "KHR_lights_punctual SpotLight outerConeAngle must be less than PI/2. Invalid light: " << light.name;
         }
